requirements.txt
 ## Install NeMo
nemo_toolkit[all]==1.21.0
librosa==0.9.1
wget
text-unidecode
matplotlib
PyYAML==6.0.1
hydra-core==1.3.1
torch
torchaudio
numpy==1.22
pyannotate
accelerate
bitsandbytes
transformers
assemblyai
whisperx @ git+https://github.com/m-bain/whisperx@main
## Install TorchAudio
